requirements.txt
boto3==1.16.28
botocore==1.19.28
bson==0.5.10
click==7.1.2
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
loguru==0.5.3
pyaml==20.4.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==5.3.1
s3transfer==0.3.3
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.2; python_version != '3.4'
